requirements.txt
astroid==2.9.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
lxml==4.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.5
pandas==1.3.5
Pillow==8.4.0
platformdirs==2.4.0
pylint==2.12.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
requests==2.26.0
six==1.16.0
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
